+++ rootdriver/llm/__init__.py
@@ -1,7 +1,13 @@
 from .llm import LLM
 from .base_adapter import BaseAdapter
+from .adapter import OpenAIAdapter, AnthropicAdapter
 
 
 
 
-__all__ = ["LLM", "BaseAdapter"]
+__all__ = [
+    "LLM",
+    "BaseAdapter",
+    "OpenAIAdapter",
+    "AnthropicAdapter",
+]